More casts

+++ src/cron.ts
@@ -120,8 +120,8 @@ async function scheduleCronJobs(
           ((json->'job')->'payload')::json as payload,
           ((json->'job')->>'queueName')::text as queue_name,
           ((json->'job')->>'runAt')::timestamptz as run_at,
-          ((json->'job')->>'maxAttempts')::int as max_attempts,
-          ((json->'job')->>'priority')::int as priority
+          ((json->'job')->>'maxAttempts')::smallint as max_attempts,
+          ((json->'job')->>'priority')::smallint as priority
         from json_array_elements($1::json) with ordinality AS entries (json, index)
       ),
       locks as (
